Fix crowd methods for binary classification

+++ maml/classifiers/_conal.py
@@ -260,7 +260,7 @@ class CoNALClassifier(MaMLClassifier):
             p_annot_pos = p_annot[..., 1]
             valid_mask = z != -1
             z[~valid_mask] = 0  # Set missing labels to 0 for loss computation, they will be ignored due to the mask.
-            loss = F.binary_cross_entropy(p_annot_pos, z, reduction="none")
+            loss = F.binary_cross_entropy(p_annot_pos, z.float(), reduction="none")
             loss = loss[valid_mask].mean()
         else:
             loss = F.cross_entropy(logits_annot.swapaxes(1, 2), z, reduction="mean", ignore_index=-1)
